its showing stuff again
